Fix a number of issues that were causing the network to not learn anything
Weights are now properly copied over to the training phase. Weights are no longer only positive, so do not explode the input. Relu is replaced with sigmoid to avoid exploding. The impage input is now correctly stored and used on both sides of WebAssembly as the 0.0 to 1.0 floating point values it always was. It is no longer erroneously treated as an u8 at any stage.