Rebase onto modern wasm-pack template
This flattens the structure and brings hot reloading for rust side changes as well. I gave up on getting CopyPlugin to play with moving mnist.js out of node_modules programatically, will want to revisit this in the future to avoid vendoring that gigantic file. Should hopefully fix a bunch of outdated packages with this rebase.